+++ src/nodejslib/bullcutter_js.cpp
@@ -20,7 +20,7 @@ BullCutterJS::BullCutterJS(const Napi::CallbackInfo &info) : Napi::ObjectWrap<Bu
 {
     Napi::Env env = info.Env();
     Napi::HandleScope scope(env);
-    int length = info.Length();
+    size_t length = info.Length();
     if (length != 3)
     {
         Napi::TypeError::New(env, "Provide 3 argument").ThrowAsJavaScriptException();
