+++ src/nodejslib/cylcutter_js.cpp
@@ -21,7 +21,7 @@ CylCutterJS::CylCutterJS(const Napi::CallbackInfo &info) : Napi::ObjectWrap<CylC
 {
     Napi::Env env = info.Env();
     Napi::HandleScope scope(env);
-    int length = info.Length();
+    size_t length = info.Length();
     if (length != 2)
     {
         Napi::TypeError::New(env, "Provide 2 argument").ThrowAsJavaScriptException();
