+++ src/nodejslib/stlreader_js.cpp
@@ -23,7 +23,7 @@ STLReaderJS::STLReaderJS(const Napi::CallbackInfo &info) : Napi::ObjectWrap<STLR
 {
     Napi::Env env = info.Env();
     Napi::HandleScope scope(env);
-    int length = info.Length();
+    size_t length = info.Length();
     if (length == 1)
     {
         Napi::TypeError::New(env, "Provide at least 2 arguments").ThrowAsJavaScriptException();
